backend/pyproject.toml
[project]
name = "trading-api"
version = "0.1.0"
description = "Trading API for executing orders with stop loss and take profit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "pytz>=2023.3",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.0.0",
    "pyjwt>=2.8.0",
    "email-validator>=2.0.0",
]

